Refactor incorporating feedback from PR#865
